feat: reorganize footer and local pages
